clean up some performance and style issues

+++ compil/src/compiler/parser/specimen_parser_unittest.cpp
@@ -2,25 +2,25 @@
 
 #include <iostream>
 
-class ParserSpecimenTests : public BaseParserTests 
+class ParserSpecimenTests : public BaseParserTests
 {
 public:
-    virtual bool checkMessage(compil::Message& expected, int mIndex)
+    virtual bool checkMessage(compil::Message& expected, size_t mIndex)
     {
         expected << compil::Message::Statement("specimen")
                  << compil::Message::Classifier("class parameter");
         return BaseParserTests::checkMessage(expected, mIndex);
     }
-    
-    bool checkSpecimen(int sIndex, int line, int column,
+
+    bool checkSpecimen(size_t sIndex, int line, int column,
                      const char* name, const char* comment = NULL)
     {
         bool result = true;
-        HF_ASSERT_LT(sIndex, (int)mDocument->objects().size());
-        
+        HF_ASSERT_LT(sIndex, mDocument->objects().size());
+
         compil::ObjectSPtr pObject = mDocument->objects()[sIndex];
         HF_ASSERT_EQ(compil::EObjectId::specimen(), pObject->runtimeObjectId());
-        compil::SpecimenSPtr pSpecimen = 
+        compil::SpecimenSPtr pSpecimen =
             boost::static_pointer_cast<compil::Specimen>(pObject);
         HF_EXPECT_STREQ(name, pSpecimen->name()->value().c_str());
         HF_EXPECT_EQ(line + 1, pSpecimen->line().value());
@@ -37,51 +37,51 @@ public:
         }
         return result;
     }
-    
-    bool checkSpecimenParameterType(int sIndex, const char* parameterType)
+
+    bool checkSpecimenParameterType(size_t sIndex, const char* parameterType)
     {
         bool result = true;
-        EXPECT_LT(sIndex, (int)mDocument->objects().size());
-        if (sIndex >= (int)mDocument->objects().size()) return false;
-        
+        EXPECT_LT(sIndex, mDocument->objects().size());
+        if (sIndex >= mDocument->objects().size()) return false;
+
         compil::ObjectSPtr pObject = mDocument->objects()[sIndex];
         EXPECT_EQ(compil::EObjectId::specimen(), pObject->runtimeObjectId());
-        compil::SpecimenSPtr pSpecimen = 
+        compil::SpecimenSPtr pSpecimen =
             boost::static_pointer_cast<compil::Specimen>(pObject);
-        
+
         compil::TypeSPtr pParameterType = pSpecimen->parameterType().lock();
         HF_ASSERT_TRUE(pParameterType);
         EXPECT_STREQ(parameterType, pParameterType->name()->value().c_str());
         return result;
     }
-    
-    bool checkConstant(int sIndex, int cIndex,
-                       int line, int column, 
+
+    bool checkConstant(size_t sIndex, size_t cIndex,
+                       int line, int column,
                        const char* name, const char* value)
     {
-        HF_ASSERT_LT(sIndex, (int)mDocument->objects().size());
+        HF_ASSERT_LT(sIndex, mDocument->objects().size());
 
         compil::ObjectSPtr sobject = mDocument->objects()[sIndex];
         HF_ASSERT_EQ(compil::EObjectId::specimen(), sobject->runtimeObjectId());
-        
+
         compil::SpecimenSPtr specimen =
             boost::static_pointer_cast<compil::Specimen>(sobject);
-        HF_ASSERT_LT(cIndex, (int)specimen->constants().size());
-        
+        HF_ASSERT_LT(cIndex, specimen->constants().size());
+
         compil::ConstantSPtr constant = specimen->constants()[cIndex];
         HF_ASSERT_EQ(line + 1, constant->line().value());
         HF_ASSERT_EQ(column, constant->column().value());
         HF_ASSERT_STREQ(name, constant->name()->value().c_str());
         HF_ASSERT_STREQ(value, constant->value().c_str());
-        
+
         return true;
     }
-    
+
 protected:
 };
 
 /*
- 
+
 specimen<string> SourceId
 {
 }
@@ -101,7 +101,7 @@ TEST_F(ParserSpecimenTests, specimenComment)
 {
     ASSERT_FALSE( parseDocument(
         "specimen //") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 10, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 12, compil::Message::p_expectStatementName));
@@ -111,7 +111,7 @@ TEST_F(ParserSpecimenTests, specimenCommentBaseTypeOpen)
 {
     ASSERT_FALSE( parseDocument(
         "specimen /* */ <") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 10, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 17, compil::Message::p_expectType));
@@ -121,7 +121,7 @@ TEST_F(ParserSpecimenTests, specimenCommentBaseTypeOpenType)
 {
     ASSERT_FALSE( parseDocument(
         "specimen < /* */ integer") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 12, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 25, compil::Message::p_expectClosingAngleBracket));
@@ -131,7 +131,7 @@ TEST_F(ParserSpecimenTests, specimenCommentBaseType)
 {
     ASSERT_FALSE( parseDocument(
         "specimen < integer /* */ >") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 20, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 27, compil::Message::p_expectStatementName));
@@ -141,7 +141,7 @@ TEST_F(ParserSpecimenTests, specimenCommentName)
 {
     ASSERT_FALSE( parseDocument(
         "specimen /* */ name") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 10, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 20, compil::Message::p_expectStatementBody));
@@ -151,7 +151,7 @@ TEST_F(ParserSpecimenTests, specimenName)
 {
     ASSERT_FALSE( parseDocument(
         "specimen name") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 1, 14, compil::Message::p_expectStatementBody));
 }
@@ -160,7 +160,7 @@ TEST_F(ParserSpecimenTests, specimenNameInherit)
 {
     ASSERT_FALSE( parseDocument(
         "specimen name inherit") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 1, 22, compil::Message::p_expectClassifierStatementName,
                                             compil::Message::Classifier("base")));
@@ -170,7 +170,7 @@ TEST_F(ParserSpecimenTests, specimenNameOpen)
 {
     ASSERT_FALSE( parseDocument(
         "specimen name {") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 1, 16, compil::Message::p_unexpectEOFInStatementBody));
 }
@@ -179,7 +179,7 @@ TEST_F(ParserSpecimenTests, specimenNameCommentOpen)
 {
     ASSERT_FALSE( parseDocument(
         "specimen name /* */ {") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 15, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 22, compil::Message::p_unexpectEOFInStatementBody));
@@ -189,7 +189,7 @@ TEST_F(ParserSpecimenTests, specimenNameOpenClose)
 {
     ASSERT_TRUE( parseDocument(
         "specimen name {}") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     EXPECT_TRUE(checkSpecimen(0, 1, 1, "name"));
     EXPECT_TRUE(checkSpecimenParameterType(0, "integer"));
@@ -199,7 +199,7 @@ TEST_F(ParserSpecimenTests, specimenNameParameterTypeOpenClose)
 {
     ASSERT_TRUE( parseDocument(
         "specimen<small> name {}") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     EXPECT_TRUE(checkSpecimen(0, 1, 1, "name"));
     EXPECT_TRUE(checkSpecimenParameterType(0, "small"));
@@ -209,7 +209,7 @@ TEST_F(ParserSpecimenTests, specimenNameMissingParameterTypeOpenClose)
 {
     ASSERT_FALSE( parseDocument(
         "specimen<blah> name {}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 1, 10, compil::Message::p_unknownClassifierType,
                   compil::Message::Type("blah")));
@@ -224,10 +224,10 @@ TEST_F(ParserSpecimenTests, 2specimensWithComments)
         "specimen name2 {}") );
 
     EXPECT_EQ(2U, mDocument->objects().size());
-    
+
     EXPECT_TRUE(checkSpecimen(0, 2, 1, "name1", "comment1"));
     EXPECT_TRUE(checkSpecimen(1, 4, 1, "name2", "comment2"));
-    
+
     EXPECT_EQ(0U, mpParser->messages().size());
 }
 
@@ -237,7 +237,7 @@ TEST_F(ParserSpecimenTests, specimenNameOpenConstant)
         "specimen name\n"
         "{\n"
         "  constant") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 3, 11, compil::Message::p_expectStatementName,
                                             compil::Message::Statement("constant")));
@@ -250,7 +250,7 @@ TEST_F(ParserSpecimenTests, specimenNameOpenBlah)
         "specimen name\n"
         "{\n"
         "  blah") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 3, 3, compil::Message::p_unknownStatment,
                                            compil::Message::Context("specimen item"),
@@ -277,7 +277,7 @@ TEST_F(ParserSpecimenTests, specimenConstantName)
         "{\n"
         "  constant cname\n"
         "}"));
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 4, 1, compil::Message::p_expectAssignmentOperator));
 }
@@ -289,7 +289,7 @@ TEST_F(ParserSpecimenTests, specimenConstantNameGarbage)
         "{\n"
         "  constant cname blah\n"
         "}"));
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 3, 18, compil::Message::p_expectAssignmentOperator));
 }
@@ -306,7 +306,6 @@ TEST_F(ParserSpecimenTests, specimenConstantNameEqual)
     EXPECT_TRUE(checkErrorMessage(0, 4, 1, compil::Message::p_expectValue,
                                            compil::Message::Statement("constant")));
 }
-
 
 TEST_F(ParserSpecimenTests, specimenConstantNameEqualSemicolon)
 {
