+++ compil/src/compiler/parser_enumration_unittest.cpp
@@ -2,24 +2,24 @@
 
 #include <iostream>
 
-class ParserEnumerationTests : public BaseParserTests 
+class ParserEnumerationTests : public BaseParserTests
 {
 public:
-    virtual bool checkMessage(compil::Message& expected, int mIndex)
+    virtual bool checkMessage(compil::Message& expected, size_t mIndex)
     {
         expected << compil::Message::Statement("enumeration")
                  << compil::Message::Classifier("class parameter");
         return BaseParserTests::checkMessage(expected, mIndex);
     }
 
-    void checkEnumeration(int eIndex, int line, int column, 
+    void checkEnumeration(size_t eIndex, int line, int column,
                      const char* name, const char* comment = NULL)
     {
-        ASSERT_LT(eIndex, (int)mDocument->objects().size());
-        
+        ASSERT_LT(eIndex, mDocument->objects().size());
+
         compil::ObjectSPtr pObject = mDocument->objects()[eIndex];
         ASSERT_EQ(compil::EObjectId::enumeration(), pObject->runtimeObjectId());
-        compil::EnumerationSPtr pEnumeration = 
+        compil::EnumerationSPtr pEnumeration =
             boost::static_pointer_cast<compil::Enumeration>(pObject);
         EXPECT_STREQ(name, pEnumeration->name()->value().c_str());
         EXPECT_EQ(lang::compil::Line(line + 1), pEnumeration->line());
@@ -35,64 +35,64 @@ public:
             ASSERT_FALSE(pEnumeration->comment());
         }
     }
-    
-    void checkEnumerationCast(int eIndex, const compil::CastableType::ECast& cast)
+
+    void checkEnumerationCast(size_t eIndex, const compil::CastableType::ECast& cast)
     {
-        ASSERT_LT(eIndex, (int)mDocument->objects().size());
-        
+        ASSERT_LT(eIndex, mDocument->objects().size());
+
         compil::ObjectSPtr pObject = mDocument->objects()[eIndex];
         ASSERT_EQ(compil::EObjectId::enumeration(), pObject->runtimeObjectId());
-        compil::EnumerationSPtr pEnumerationt = 
+        compil::EnumerationSPtr pEnumerationt =
             boost::static_pointer_cast<compil::Enumeration>(pObject);
 
         EXPECT_EQ(cast, pEnumerationt->cast());
     }
-    
-    void checkEnumerationFlags(int eIndex, bool flags)
+
+    void checkEnumerationFlags(size_t eIndex, bool flags)
     {
-        ASSERT_LT(eIndex, (int)mDocument->objects().size());
-        
+        ASSERT_LT(eIndex, mDocument->objects().size());
+
         compil::ObjectSPtr pObject = mDocument->objects()[eIndex];
         ASSERT_EQ(compil::EObjectId::enumeration(), pObject->runtimeObjectId());
-        compil::EnumerationSPtr pEnumerationt = 
+        compil::EnumerationSPtr pEnumerationt =
             boost::static_pointer_cast<compil::Enumeration>(pObject);
 
         EXPECT_EQ(flags, pEnumerationt->flags());
     }
-    
-    bool checkEnumerationParameterType(int eIndex, const char* baseType)
+
+    bool checkEnumerationParameterType(size_t eIndex, const char* baseType)
     {
         bool result = true;
-        EXPECT_LT(eIndex, (int)mDocument->objects().size());
-        if (eIndex >= (int)mDocument->objects().size()) return false;
-        
+        EXPECT_LT(eIndex, mDocument->objects().size());
+        if (eIndex >= mDocument->objects().size()) return false;
+
         compil::ObjectSPtr pObject = mDocument->objects()[eIndex];
         EXPECT_EQ(compil::EObjectId::enumeration(), pObject->runtimeObjectId());
-        compil::EnumerationSPtr pEnumerationt = 
+        compil::EnumerationSPtr pEnumerationt =
         boost::static_pointer_cast<compil::Enumeration>(pObject);
-        
+
         compil::TypeSPtr pParameterType = pEnumerationt->parameterType().lock();
         HF_ASSERT_TRUE(pParameterType);
         EXPECT_STREQ(baseType, pParameterType->name()->value().c_str());
         return result;
     }
-    
-    bool checkEnumerationValue(int eIndex, int vIndex, 
-                    int line, int column, 
-                    const char* name, 
+
+    bool checkEnumerationValue(size_t eIndex, size_t vIndex,
+                    int line, int column,
+                    const char* name,
                     const char* comment = NULL)
     {
         bool result = true;
-        HF_ASSERT_LT(eIndex, (int)mDocument->objects().size());
-        
+        HF_ASSERT_LT(eIndex, mDocument->objects().size());
+
         compil::ObjectSPtr pObject = mDocument->objects()[eIndex];
         HF_ASSERT_EQ(compil::EObjectId::enumeration(), pObject->runtimeObjectId());
-        compil::EnumerationSPtr pEnumeration = 
+        compil::EnumerationSPtr pEnumeration =
             boost::static_pointer_cast<compil::Enumeration>(pObject);
-        
-        HF_ASSERT_LT(vIndex, (int)pEnumeration->enumerationValues().size());
+
+        HF_ASSERT_LT(vIndex, pEnumeration->enumerationValues().size());
         compil::ObjectSPtr pVObject = pEnumeration->enumerationValues()[vIndex];
-        compil::EnumerationValueSPtr pEnumerationValue = 
+        compil::EnumerationValueSPtr pEnumerationValue =
             boost::static_pointer_cast<compil::EnumerationValue>(pVObject);
         HF_EXPECT_EQ(lang::compil::Line(line + 1), pEnumerationValue->line());
         HF_EXPECT_EQ(lang::compil::Column(column), pEnumerationValue->column());
@@ -109,12 +109,12 @@ public:
         }
         return result;
     }
-    
+
 protected:
 };
 
 /*
- 
+
 enum Weekday
 {
     sunday;
@@ -141,7 +141,7 @@ TEST_F(ParserEnumerationTests, enumComment)
 {
     ASSERT_FALSE( parseDocument(
         "enum //") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 6, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 8, compil::Message::p_expectStatementName));
@@ -151,7 +151,7 @@ TEST_F(ParserEnumerationTests, enumCommentParameterTypeOpen)
 {
     ASSERT_FALSE( parseDocument(
         "enum /* */ <") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 6, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 13, compil::Message::p_expectType));
@@ -161,7 +161,7 @@ TEST_F(ParserEnumerationTests, enumCommentParameterTypeOpenType)
 {
     ASSERT_FALSE( parseDocument(
         "enum < /* */ integer") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 8, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 21, compil::Message::p_expectClosingAngleBracket));
@@ -171,7 +171,7 @@ TEST_F(ParserEnumerationTests, enumCommentParameterType)
 {
     ASSERT_FALSE( parseDocument(
         "enum < integer /* */ >") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 16, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 23, compil::Message::p_expectStatementName));
@@ -181,7 +181,7 @@ TEST_F(ParserEnumerationTests, enumCommentName)
 {
     ASSERT_FALSE( parseDocument(
         "enum /* */ name") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 6, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 16, compil::Message::p_expectStatementBody));
@@ -191,7 +191,7 @@ TEST_F(ParserEnumerationTests, enumName)
 {
     ASSERT_FALSE( parseDocument(
         "enum name") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 1, 10, compil::Message::p_expectStatementBody));
 }
@@ -200,7 +200,7 @@ TEST_F(ParserEnumerationTests, enumNameOpen)
 {
     ASSERT_FALSE( parseDocument(
         "enum name {") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 1, 12, compil::Message::p_unexpectEOFInStatementBody));
 }
@@ -209,7 +209,7 @@ TEST_F(ParserEnumerationTests, enumNameCommentOpen)
 {
     ASSERT_FALSE( parseDocument(
         "enum name /* */ {") );
-    
+
     ASSERT_EQ(2U, mpParser->messages().size());
     EXPECT_TRUE(checkWarningMessage(0, 1, 11, compil::Message::p_misplacedComment));
     EXPECT_TRUE(checkErrorMessage(1, 1, 18, compil::Message::p_unexpectEOFInStatementBody));
@@ -219,7 +219,7 @@ TEST_F(ParserEnumerationTests, enumNameOpenClose)
 {
     ASSERT_TRUE( parseDocument(
         "enum name {}") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     checkEnumeration(0, 1, 1, "name");
     checkEnumerationCast(0, compil::CastableType::ECast::weak());
@@ -231,7 +231,7 @@ TEST_F(ParserEnumerationTests, enumNameParameterTypeOpenClose)
 {
     ASSERT_TRUE( parseDocument(
         "enum<small> name {}") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     checkEnumeration(0, 1, 1, "name");
     checkEnumerationCast(0, compil::CastableType::ECast::weak());
@@ -243,7 +243,7 @@ TEST_F(ParserEnumerationTests, enumNameWrongParameterTypeOpenClose)
 {
     ASSERT_FALSE( parseDocument(
         "enum<real32> name {}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 1, 1, compil::Message::v_unacceptableParameterType,
                 compil::Message::Options("small, short or integer")));
@@ -253,7 +253,7 @@ TEST_F(ParserEnumerationTests, enumNameMissingParameterTypeOpenClose)
 {
     ASSERT_FALSE( parseDocument(
         "enum<blah> name {}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 1, 6, compil::Message::p_unknownClassifierType,
                   compil::Message::Type("blah")));
@@ -263,7 +263,7 @@ TEST_F(ParserEnumerationTests, weakEnumNameCommentOpenClose)
 {
     ASSERT_TRUE( parseDocument(
         "weak enum name {}") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     checkEnumeration(0, 1, 1, "name");
     checkEnumerationFlags(0, false);
@@ -274,7 +274,7 @@ TEST_F(ParserEnumerationTests, strongEnumNameCommentOpenClose)
 {
     ASSERT_TRUE( parseDocument(
         "strong enum name {}") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     checkEnumeration(0, 1, 1, "name");
     checkEnumerationFlags(0, false);
@@ -285,7 +285,7 @@ TEST_F(ParserEnumerationTests, flagsEnumNameCommentOpenClose)
 {
     ASSERT_TRUE( parseDocument(
         "flags enum name {}") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     checkEnumeration(0, 1, 1, "name");
     checkEnumerationFlags(0, true);
@@ -301,10 +301,10 @@ TEST_F(ParserEnumerationTests, 2enumsWithComments)
         "enum name2 {}") );
 
     EXPECT_EQ(2U, mDocument->objects().size());
-    
+
     checkEnumeration(0, 2, 1, "name1", "comment1");
     checkEnumeration(1, 4, 1, "name2", "comment2");
-    
+
     EXPECT_EQ(0U, mpParser->messages().size());
 }
 
@@ -315,7 +315,7 @@ TEST_F(ParserEnumerationTests, enumValue)
         "{\n"
         "  value\n"
         "}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 4, 1, compil::Message::p_expectSemicolon));
 }
@@ -327,7 +327,7 @@ TEST_F(ParserEnumerationTests, enumValueSomething)
         "{\n"
         "  value blah\n"
         "}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 3, 9, compil::Message::p_expectSemicolon));
 }
@@ -339,7 +339,7 @@ TEST_F(ParserEnumerationTests, enumValueEqual)
         "{\n"
         "  value =\n"
         "}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 4, 1, compil::Message::p_unexpectEOFInStatementBody));
 }
@@ -351,7 +351,7 @@ TEST_F(ParserEnumerationTests, enumValueEqualValue)
         "{\n"
         "  value = blah\n"
         "}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 3, 11, compil::Message::p_expectStatementName,
                                   compil::Message::Statement("enumeration value")));
@@ -365,7 +365,7 @@ TEST_F(ParserEnumerationTests, enumValue1EqualValue2EqualValue1)
         "  value1;\n"
         "  value2 = value1\n"
         "}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 5, 1, compil::Message::p_expectSemicolon));
 }
@@ -378,11 +378,10 @@ TEST_F(ParserEnumerationTests, enumValue1EqualValue2EqualValue1Or)
         "  value1;\n"
         "  value2 = value1 |\n"
         "}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 5, 1, compil::Message::p_unexpectEOFInStatementBody));
 }
-
 
 TEST_F(ParserEnumerationTests, enumValue1EqualValue2EqualValue1OrSemicolon)
 {
@@ -392,7 +391,7 @@ TEST_F(ParserEnumerationTests, enumValue1EqualValue2EqualValue1OrSemicolon)
         "  value1;\n"
         "  value2 = value1 |;\n"
         "}") );
-    
+
     ASSERT_EQ(1U, mpParser->messages().size());
     EXPECT_TRUE(checkErrorMessage(0, 4, 20, compil::Message::p_unexpectEOFInStatementBody));
 }
@@ -439,7 +438,6 @@ TEST_F(ParserEnumerationTests, enumComposedValue)
     EXPECT_TRUE(checkEnumerationValue(0, 2, 5, 3, "value3"));
 }
 
-
 TEST_F(ParserEnumerationTests, enum2ValueWithComment)
 {
     ASSERT_TRUE( parseDocument(
@@ -452,4 +450,3 @@ TEST_F(ParserEnumerationTests, enum2ValueWithComment)
     checkEnumeration(0, 1, 1, "name");
     EXPECT_TRUE(checkEnumerationValue(0, 0, 4, 3, "value", "comment"));
 }
-
