+++ compil/src/compiler/parser_inner_enumration_unittest.cpp
@@ -2,27 +2,27 @@
 
 #include <iostream>
 
-class ParserInnerEnumerationTests : public BaseParserTests 
+class ParserInnerEnumerationTests : public BaseParserTests
 {
 public:
-    bool checkEnumeration(int eIndex, int line, int column, 
+    bool checkEnumeration(size_t eIndex, int line, int column,
                      const char* name, const char* comment = NULL)
     {
         bool result = true;
-    
-        HF_ASSERT_LT(0, (int)mDocument->objects().size());
-        
+
+        HF_ASSERT_LT(0U, mDocument->objects().size());
+
         compil::ObjectSPtr pSObject = mDocument->objects()[0];
         HF_EXPECT_EQ(compil::EObjectId::structure(), pSObject->runtimeObjectId());
-        compil::StructureSPtr pStructure = 
+        compil::StructureSPtr pStructure =
             boost::static_pointer_cast<compil::Structure>(pSObject);
-            
-        HF_ASSERT_LT(eIndex, (int)pStructure->objects().size());
+
+        HF_ASSERT_LT(eIndex, pStructure->objects().size());
         compil::ObjectSPtr pEObject = pStructure->objects()[eIndex];
         HF_EXPECT_EQ(compil::EObjectId::enumeration(), pEObject->runtimeObjectId());
-        compil::EnumerationSPtr pEnumeration = 
+        compil::EnumerationSPtr pEnumeration =
             boost::static_pointer_cast<compil::Enumeration>(pEObject);
-            
+
         EXPECT_STREQ(name, pEnumeration->name()->value().c_str());
         HF_EXPECT_EQ(lang::compil::Line(line + 1), pEnumeration->line());
         HF_EXPECT_EQ(lang::compil::Column(column), pEnumeration->column());
@@ -39,7 +39,7 @@ public:
 
         return result;
     }
-    
+
 protected:
 };
 
@@ -47,7 +47,7 @@ TEST_F(ParserInnerEnumerationTests, enum)
 {
     ASSERT_TRUE( parseDocument(
         "structure sname { enum ename {} }") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     EXPECT_TRUE(checkEnumeration(0, 1, 19, "ename"));
 }
@@ -56,7 +56,7 @@ TEST_F(ParserInnerEnumerationTests, strongEnum)
 {
     ASSERT_TRUE( parseDocument(
         "structure sname { strong enum ename {} }") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     EXPECT_TRUE(checkEnumeration(0, 1, 19, "ename"));
 }
@@ -65,7 +65,7 @@ TEST_F(ParserInnerEnumerationTests, weakEnum)
 {
     ASSERT_TRUE( parseDocument(
         "structure sname { weak enum ename {} }") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     EXPECT_TRUE(checkEnumeration(0, 1, 19, "ename"));
 }
@@ -74,20 +74,18 @@ TEST_F(ParserInnerEnumerationTests, 2Enums)
 {
     ASSERT_TRUE( parseDocument(
         "structure sname { enum ename1 {} enum ename2 {} }") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     EXPECT_TRUE(checkEnumeration(0, 1, 19, "ename1"));
     EXPECT_TRUE(checkEnumeration(1, 1, 34, "ename2"));
 }
 
-
 TEST_F(ParserInnerEnumerationTests, 2EnumsWithComments)
 {
     ASSERT_TRUE( parseDocument(
         "structure sname { /*comment 1*/ enum ename1 {} /*comment 2*/ enum ename2 {} }") );
-    
+
     EXPECT_EQ(1U, mDocument->objects().size());
     EXPECT_TRUE(checkEnumeration(0, 1, 33, "ename1", "comment 1"));
     EXPECT_TRUE(checkEnumeration(1, 1, 62, "ename2", "comment 2"));
 }
-
